Move OrcScan into shims
